requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
authlib>=1.2.0
python-dotenv>=0.21.0
requests>=2.31.0
httpx>=0.24.0
itsdangerous>=2.1.2
pandas>=2.2.1
streamlit>=1.25.0
openpyxl>=3.1.2
plotly>=5.15.0
matplotlib>=3.9,<3.11
xlrd>=2.0.1          
supabase>=2.5.1
storage3>=0.7.4
psutil>=7.1.3